[klibc] Remove unused no-MMU support
There is some conditional logic to handle no-MMU architectures, but no
such architectures are defined.  We are now having to override the
automatic detection for each new architecture because it only
checks for fork() and not clone().

Since this is currently unused and therefore untested, it's probably
broken.  Remove it.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>