[klibc] mips64: Set executable base address to fix link failure on R6
It appears that at some point the GNU tools started using a different
default base address for MIPS64 R6 executables; at least this is what
I see with the Debian-provided cross-toolchain.  This resulted in
klibc.so being too far away from the executable for regular jumps to
work.

Add an explicit linker option to set the base address to the value
we expect.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>