[klibc] mount_opts: Fix the mount_opts str length
Newer versions of LLVM report:

usr/utils/mount_opts.c:20:3: error: initializer-string for character array is too long, array size is 8 but initializer has size 9 (including the null terminating character); did you mean to use the 'nonstring' attribute? [-Werror,-Wunterminated-string-initialization]
   20 |         {"diratime", MS_NODIRATIME, 0, MS_NODIRATIME},
      |          ^~~~~~~~~~

This is indeed a bit odd. "diratime" is 9 bytes long with the \0
terminator but placed into a struct that uses a static length of 8 bytes
for that buffer. I suppose this can cause all sorts of undefined
behaviors in theory but that in practice this never caused anything bad
because the next field is an unsigned long containing 2048 so the upper
bytes are 0 and act as string terminators by chance.

Anyway, fixing this helps unblock builds with newer toolchains.

Signed-off-by: Florent Revest <revest@chromium.org>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>